-change primary key identifier from Long to Integer Insert counter entry to Serial Number table if it does not exisit

+++ src/main/java/com/ipayafrica/elipapower/service/ITokenRequestService.java
@@ -8,9 +8,9 @@ import org.springframework.data.domain.Pageable;
 
 import com.ipayafrica.elipapower.model.TokenRequest;
 
-public interface ITokenRequestService extends IGenericService<TokenRequest,Long>{
+public interface ITokenRequestService extends IGenericService<TokenRequest,Integer>{
 	List<TokenRequest> getAllRecords();
-	TokenRequest getByTokenRequestId (Long requestId);
+	TokenRequest getByTokenRequestId (Integer requestId);
 	TokenRequest findTokenRequestByRef(Double ref);
 	TokenRequest findTokenRequestByMeterNo(String meterno);
 	List<TokenRequest> findFailedRequests(Byte status, Date requestdate);
